Fixed issue in MeterManager.java
- Added check in pushMeterMetrics method to delete Meter only when it is in the PENDING_REMOVE state
- Minor change: Corrected description in MeterId.java to indicate that MeterIds are unique only within a device

+++ incubator/net/src/main/java/org/onosproject/incubator/net/meter/impl/MeterManager.java
@@ -222,7 +222,7 @@ public class MeterManager extends AbstractListenerProviderRegistry<MeterEvent, M
                     provider().performMeterOperation(m.deviceId(),
                                                      new MeterOperation(m,
                                                                         MeterOperation.Type.ADD));
-                } else {
+                } else if ((m.state() == MeterState.PENDING_REMOVE)) {
                     store.deleteMeterNow(m);
                 }
             });
